+++ spec/rails_state_machine/model_spec.rb
@@ -43,6 +43,46 @@ describe RailsStateMachine::Model do
       )
     end
 
+    it 'defines state constants with a prefix if a prefix is given' do
+      class_with_machine = Class.new(ActiveRecord::Base) do
+        include RailsStateMachine::Model
+        state_machine(:foo, prefix: 'prefix_foo') do
+          state :completed
+        end
+
+        state_machine(:bar, prefix: 'prefix_bar') do
+          state :completed
+        end
+      end
+
+      expect(class_with_machine::PREFIX_FOO_STATE_COMPLETED).to be_present
+      expect(class_with_machine::PREFIX_BAR_STATE_COMPLETED).to be_present
+      expect{ class_with_machine::STATE_COMPLETED }.to raise_error(NameError)
+    end
+
+    it 'defines state methods with a prefix if a prefix is given' do
+      parcel = parcel_class.new
+      expect(parcel).to respond_to(:review_draft?)
+      expect(parcel).to respond_to(:review_approved?)
+      expect(parcel).to_not respond_to(:draft?)
+      expect(parcel).to_not respond_to(:approved?)
+    end
+
+    it 'throws an exception if multiple state machines on the same model define the same state name' do
+      expect {
+        class_with_machine = Class.new(ActiveRecord::Base) do
+          include RailsStateMachine::Model
+          state_machine(:foo) do
+            state :completed
+          end
+
+          state_machine(:bar) do
+            state :completed
+          end
+        end
+      }.to raise_error(RailsStateMachine::StateMachine::StateAlreadyDefinedError, 'State :completed has already been defined in the :foo state machine. You may use the :prefix option when defining a state machine to avoid that.')
+    end
+
   end
 
   it 'does not remove already defined state machines if you include the module again afterwards (BUGFIX)' do