+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added: State machine can now use the `:prefix` option to avoid name collision if you define multiple state machines 
+  on the same model, and use state names more than once
 
 ### Compatible changes